# tutorial/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InteractionData": {
        "additionalProperties": false,
        "description": "Student interaction payload.",
        "properties": {
          "interaction": {
            "additionalProperties": true,
            "title": "Interaction",
            "type": "object"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          }
        },
        "required": [
          "student_id",
          "interaction"
        ],
        "title": "InteractionData",
        "type": "object"
      },
...
      "LessonDetail": {
        "additionalProperties": false,
        "description": "Full lesson payload.",
        "properties": {
          "concept_coverage": {
            "additionalProperties": true,
            "title": "Concept Coverage",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "csta_standards": {
            "items": {},
            "title": "Csta Standards",
            "type": "array"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "estimated_duration_minutes": {
            "title": "Estimated Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "interactive_elements": {
            "items": {},
            "title": "Interactive Elements",
            "type": "array"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "teaching_effectiveness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teaching Effectiveness Score"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "incident_id",
          "title",
          "narrative",
          "interactive_elements",
          "difficulty",
          "csta_standards",
          "estimated_duration_minutes",
          "concept_coverage",
          "teaching_effectiveness_score",
          "created_at",
          "updated_at"
        ],
        "title": "LessonDetail",
        "type": "object"
      },
      "LessonGenerateRequest": {
        "additionalProperties": false,
        "description": "Request body for generating a lesson from defense outcomes.",
        "properties": {
          "experience_level": {
            "default": "intermediate",
            "maxLength": 100,
            "minLength": 1,
            "title": "Experience Level",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "preferred_learning_style": {
            "default": "visual",
            "maxLength": 100,
            "minLength": 1,
            "title": "Preferred Learning Style",
            "type": "string"
          },
          "student_name": {
            "default": "learner",
            "maxLength": 200,
            "minLength": 1,
            "title": "Student Name",
            "type": "string"
          }
        },
        "required": [
          "incident_id"
        ],
        "title": "LessonGenerateRequest",
        "type": "object"
      },
      "LessonGenerateResponse": {
        "additionalProperties": false,
        "description": "Teaching ticket summary.",
        "properties": {
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "lesson_id": {
            "format": "uuid",
            "title": "Lesson Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "teaching_thread_id": {
            "title": "Teaching Thread Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "lesson_id",
          "teaching_thread_id",
          "status"
        ],
        "title": "LessonGenerateResponse",
        "type": "object"
      },
...
      "LessonResponse": {
        "additionalProperties": false,
        "description": "Lesson list item.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "estimated_duration_minutes": {
            "title": "Estimated Duration Minutes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "incident_id",
          "title",
          "narrative",
          "difficulty",
          "estimated_duration_minutes",
          "created_at"
        ],
        "title": "LessonResponse",
        "type": "object"
      },
...
      "SandboxInfoResponse": {
        "additionalProperties": false,
        "description": "Sandbox metadata for a lesson.",
        "properties": {
          "lesson_id": {
            "format": "uuid",
            "title": "Lesson Id",
            "type": "string"
          },
          "resources": {
            "additionalProperties": true,
            "title": "Resources",
            "type": "object"
          },
          "sandbox_mode": {
            "default": "docker",
            "title": "Sandbox Mode",
            "type": "string"
          },
          "workspace_hint": {
            "default": "Use the teaching sandbox agent for isolated execution.",
            "title": "Workspace Hint",
            "type": "string"
          }
        },
        "required": [
          "lesson_id"
        ],
        "title": "SandboxInfoResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The World's First Agentic AI That Learns by Teaching",
    "title": "TUTORIAL API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/lessons/": {
      "get": {
        "description": "List persisted lessons, optionally filtered by difficulty.",
        "operationId": "list_lessons_api_v1_lessons__get",
        "parameters": [
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LessonResponse"
                  },
                  "title": "Response List Lessons Api V1 Lessons  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lessons",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/v1/lessons/generate": {
      "post": {
        "description": "Generate an interactive lesson anchored to a defended incident.",
        "operationId": "generate_lesson_api_v1_lessons_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Lesson",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/v1/lessons/{lesson_id}": {
      "get": {
        "description": "Return lesson narrative and interactive elements.",
        "operationId": "get_lesson_api_v1_lessons__lesson_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lesson",
        "tags": [
          "lessons"
        ]
      }
    },
...
    "/api/v1/lessons/{lesson_id}/interaction": {
      "post": {
        "description": "Append a structured interaction to the student's progress row.",
        "operationId": "record_interaction_api_v1_lessons__lesson_id__interaction_post",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Interaction",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/v1/lessons/{lesson_id}/sandbox": {
      "get": {
        "description": "Describe sandbox expectations for the lesson.",
        "operationId": "get_sandbox_api_v1_lessons__lesson_id__sandbox_get",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sandbox",
        "tags": [
          "lessons"
        ]
      }
    },
    "/api/v1/lessons/{lesson_id}/start": {
      "post": {
        "description": "Initialize progress tracking for a student.",
        "operationId": "start_lesson_api_v1_lessons__lesson_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "student_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Student Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Start Lesson Api V1 Lessons  Lesson Id  Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Lesson",
        "tags": [
          "lessons"
        ]
      }
    },
...
    "/api/v1/system/flywheel": {
      "get": {
        "description": "Orchestration knowledge graph stats and flywheel signals (defense \u2192 teaching loop).",
        "operationId": "flywheel_snapshot_api_v1_system_flywheel_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Flywheel Snapshot Api V1 System Flywheel Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Flywheel Snapshot",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/v1/system/shutdown": {
      "post": {
        "description": "Request a graceful coordinator shutdown.",
        "operationId": "shutdown_api_v1_system_shutdown_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Shutdown Api V1 System Shutdown Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shutdown",
        "tags": [
          "system"
        ]
      }
    }
  }
}